Rename, clean code, comment

+++ c++/nda/indexmap/permutation.hpp
@@ -11,13 +11,13 @@ namespace nda::permutations {
   // to std::array<int, Rank>
   // indeed, we can not yet (C++20 ?) template array, array_view on a std::array ...
 
-  template <int Rank> constexpr std::array<int, Rank> decode(uint64_t binary_representation) {
+  template <size_t Rank> constexpr std::array<int, Rank> decode(uint64_t binary_representation) {
     auto result = nda::make_initialized_array<Rank>(0);
     for (int i = 0; i < Rank; ++i) result[i] = (binary_representation >> (4 * i)) & 0b1111ull;
     return result;
   }
 
-  template <int Rank> constexpr uint64_t encode(std::array<int, Rank> const &permutation) {
+  template <size_t Rank> constexpr uint64_t encode(std::array<int, Rank> const &permutation) {
     uint64_t result = 0;
     for (int i = 0; i < Rank; ++i) result += (permutation[i] >> (4 * i)) & 0b1111ull;
     return result;
